# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Shift Sensei (シフト先生) - a fast monthly shift-scheduling tool for nursing home unit leaders
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List all staff members
      responses:
        "200":
          description: List of staff members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Staff"
    post:
      operationId: createStaff
      tags: [staff]
      summary: Create a staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffInput"
      responses:
        "201":
          description: Created staff member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"
  /staff/reorder:
    post:
      operationId: reorderStaff
      tags: [staff]
      summary: Set the display order of staff, which also drives the schedule grid's row order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffReorderBody"
      responses:
        "200":
          description: Staff members in their new order
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Staff"
  /staff/{id}:
    patch:
      operationId: updateStaff
      tags: [staff]
      summary: Update a staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffUpdate"
      responses:
        "200":
          description: Updated staff member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteStaff
      tags: [staff]
      summary: Delete a staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /shifts:
    get:
      operationId: listShifts
      tags: [shifts]
      summary: List shifts, optionally filtered by month and/or staff member
      parameters:
        - name: month
          in: query
          required: false
          description: "Filter by month in YYYY-MM format"
          schema:
            type: string
        - name: staffId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of shifts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shift"
    post:
      operationId: createShift
      tags: [shifts]
      summary: Create a scheduled shift
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShiftInput"
      responses:
        "201":
          description: Created shift
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
  /shifts/{id}:
    patch:
      operationId: updateShift
      tags: [shifts]
      summary: Update a scheduled shift
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShiftUpdate"
      responses:
        "200":
          description: Updated shift
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteShift
      tags: [shifts]
      summary: Delete a scheduled shift
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /shifts/reset-month:
    post:
      operationId: resetShiftMonth
      tags: [shifts]
      summary: >
        Delete all non-constraint shifts for a month.
        Preserves hope_off, paid_leave, and attendance constraints,
        night-duty ownership marks, and day remarks.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetShiftMonthInput"
      responses:
        "200":
          description: Reset complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResetShiftMonthResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Staff:
      type: object
      required:
        [
          id,
          name,
          color,
          shortLabel,
          isLeader,
          availableShiftTypeIds,
          nightShiftAvailable,
          maxShifts,
          sortOrder,
          createdAt,
        ]
      properties:
        id:
          type: integer
        name:
          type: string
        shortLabel:
          type: string
          description: "1-3 character label rendered on the schedule grid's staff row"
        color:
          type: string
          description: "Hex color used to identify this staff member in the shift table"
        isLeader:
          type: boolean
          description: "Marks a unit leader. Storage only; not yet used elsewhere in the app"
        availableShiftTypeIds:
          description: >
            Reserved for a future "which shift types this staff member can be
            assigned" restriction. Null/empty means no restriction. Not enforced
            anywhere yet.
          oneOf:
            - type: array
              items:
                type: integer
            - type: "null"
        nightShiftAvailable:
          type: boolean
          description: "Reserved for a future night-shift eligibility rule."
        maxShifts:
          description: "Reserved for a future per-staff monthly shift-count cap."
          oneOf:
            - type: integer
            - type: "null"
        sortOrder:
          type: integer
          description: "Drives staff row order in Staff Management and on the schedule grid"
        createdAt:
          type: string
          format: date-time
    StaffInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        shortLabel:
          type: string
          minLength: 1
          maxLength: 3
        color:
          type: string
        isLeader:
          type: boolean
        availableShiftTypeIds:
          type: array
          items:
            type: integer
        nightShiftAvailable:
          type: boolean
        maxShifts:
          oneOf:
            - type: integer
            - type: "null"
        sortOrder:
          type: integer
    StaffUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        shortLabel:
          type: string
          minLength: 1
          maxLength: 3
        color:
          type: string
        isLeader:
          type: boolean
        availableShiftTypeIds:
          oneOf:
            - type: array
              items:
                type: integer
            - type: "null"
        nightShiftAvailable:
          type: boolean
        maxShifts:
          oneOf:
            - type: integer
            - type: "null"
        sortOrder:
          type: integer
    StaffReorderBody:
      type: object
      description: "Full list of staff ids in their new display/row order"
      required: [ids]
      properties:
        ids:
          type: array
          items:
            type: integer
          minItems: 1
          description: "Staff ids in the desired order"
    ShiftCode:
      type: string
      minLength: 1
      description: >
        Shift stamp code placed on the shift table for a given staff member
        and date. References a ShiftType's `code` by value — the set of
        valid codes is editable in Settings ("シフト種別"), not a fixed enum.
...
    ResetShiftMonthInput:
      type: object
      required: [month]
      properties:
        month:
          type: string
          description: "Month to reset in YYYY-MM format"
    ResetShiftMonthResult:
      type: object
      required: [deletedCount]
      properties:
        deletedCount:
          type: integer
          description: Number of shift records deleted
...
    ConstraintType:
      type: string
      description: >
        A protected scheduling constraint entered in the dedicated
        "制約入力" phase, before normal shifts are assigned:
        "hope_off" (休み固定) forces code to the "holiday"-category shift
        type's code (originally "休み") and fully locks the cell (immutable
        except via the one-tap remove gesture); "paid_leave" (有給固定) is
        the same, but forces code to the "paid_leave"-category shift type's
        code (originally "有休"); "attendance" (📌 出勤) locks the cell
        against deletion but leaves code null until a normal shift is
        assigned on top of it later.
      enum: [hope_off, attendance, paid_leave]
    Shift:
      type: object
      required: [id, staffId, date, code, constraintType, locked, createdAt]
      properties:
        id:
          type: integer
        staffId:
          type: integer
        date:
          type: string
          description: "Date in YYYY-MM-DD format"
        code:
          description: "Null for a blank 出勤 (attendance) constraint cell with no shift assigned yet"
          oneOf:
            - $ref: "#/components/schemas/ShiftCode"
            - type: "null"
        constraintType:
          description: "Null when this cell has no constraint applied"
          oneOf:
            - $ref: "#/components/schemas/ConstraintType"
            - type: "null"
        locked:
          type: boolean
          description: "Locked cells cannot be overwritten or cleared by tapping until unlocked"
        createdAt:
          type: string
          format: date-time
    ShiftInput:
      type: object
      description: >
        Either `code` or `constraintType` must be provided. When
        `constraintType` is set, the server derives `code` and `locked`
        deterministically and ignores any client-supplied values for them.
      required: [staffId, date]
      properties:
        staffId:
          type: integer
        date:
          type: string
        code:
          $ref: "#/components/schemas/ShiftCode"
        constraintType:
          $ref: "#/components/schemas/ConstraintType"
    ShiftUpdate:
      type: object
      description: >
        At least one of code, constraintType, or locked must be provided.
        Set constraintType to null to remove an existing constraint (and
        its automatic lock).
      properties:
        code:
          description: "Set to null to clear an assigned shift while keeping an attendance constraint's marker"
          oneOf:
            - $ref: "#/components/schemas/ShiftCode"
            - type: "null"
        constraintType:
          oneOf:
            - $ref: "#/components/schemas/ConstraintType"
            - type: "null"
        locked:
          type: boolean
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...